LayoutTests/fast/backgrounds/background-repeat-x-y-parse-expected.txt

This test checks that background-repeat-x/y are not parsed or supported

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.getComputedStyle(document.getElementById('test1')).getPropertyValue('background-repeat') is "repeat"
PASS window.getComputedStyle(document.getElementById('test2')).getPropertyValue('background-repeat') is "repeat"
PASS window.getComputedStyle(document.getElementById('test3')).getPropertyValue('background-repeat') is "repeat-y"
PASS window.getComputedStyle(document.getElementById('test4')).getPropertyValue('background-repeat') is "repeat-x"
PASS CSS.supports('background-repeat-x: inherit') is false
PASS CSS.supports('background-repeat-x', 'inherit') is false
PASS CSS.supports('background-repeat-y: inherit') is false
PASS CSS.supports('background-repeat-y', 'inherit') is false
PASS successfullyParsed is true

TEST COMPLETE
